requirements.txt
pyyaml
PyOpenGL
PyOpenGL_accelerate
numpy
glfw
mysql-connector-python
